Sphinx_to_vivid added
Added the sphinxToVivid.py to the sphinxval/utils/ folder and added
support for the code into the main pipeline as part of the
sphinxval/sphinx/sphinx.py code following the intiuitive_validation
function call. Currently, all this does is add a new output
file to the outputs folder called SphinxToVivid_database.json, which
would need to be imported to VIVID (and name changed to just database).
That is a future task, however.