Commit for AWTPE
Motivated by Leila's request:
"the warning time in particular is the surprise,
but I think that's because of the time diff between the flare
and threshold crossing
a percent error of this time could be interesting..
(we do that for CME arrival time)
AWT percent error
I would say AWT percent error = AWT (min) / T_trigger to threshold
crossing time (min)
(could also be trigger to onset crossing time)
so for this one Threshold crossing time 100 MeV - Flare time.
You will need to decide if its the flare peak time or threshold
crossing time "

This commit makes the changes to validation.py to support this new
metric. Follows a similar structure to the other awt metrics which
don't actually use metrics.py, but is all done in validation.py.
Adds new columns to awt_metrics for AWTPE following for the mean
AWTPE for each model that has a prediction in for all clear,
SEP start time, or SEP threshold crossing. In the awt_selections
files new columns are added to show this math done for each sep event
in the data set for each model. Will follow up on this commit with
updates to the validation unit test.